🔧 refactor: finalize clean architecture with reusable hooks and modular structure
