+++ features/auth/index.ts
@@ -1,3 +1,4 @@
+import EmailPasswordAuth from "./components/LoginForm";
 import SignUpForm from "./components/SignUpForm";
 
-export default { SignUpForm };
+export default { SignUpForm, EmailPasswordAuth };
