+++ features/auth/schema/login.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const signInSchema = z.object({
+  email: z.string().email("Please enter a valid email address"),
+  password: z.string().min(8, "Password must be at least 8 characters"),
+});
+
+export type signInValues = z.infer<typeof signInSchema>;
