+++ package.json
@@ -48,7 +48,8 @@
     "react-redux": "^9.2.0",
     "redux-persist": "^6.0.0",
     "sonner-native": "^0.19.0",
-    "tailwindcss": "3.4.17"
+    "tailwindcss": "3.4.17",
+    "expo-dev-client": "~5.1.8"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
